Fix definition to match prototype (int vs size_t)

+++ sources/lib/run-time/heap-trail.c
@@ -335,7 +335,7 @@ static void add_target_object (mps_addr_t obj)
 
 static void add_target_object_of_wrapper (mps_addr_t object, 
                                           mps_fmt_t format, mps_pool_t pool, 
-                                          void *wrapper, int p2)
+                                          void *wrapper, size_t p2)
 {
   void **found = (void**)object;
   void *found_wrapper = *found;
@@ -411,7 +411,7 @@ static BOOL trace_reference(mps_addr_t object, mps_addr_t parent,
 
 static void look_for_reference(mps_addr_t object, mps_fmt_t format, 
                                mps_pool_t pool, 
-                               void *p1, int p2)
+                               void *p1, size_t p2)
 {
   trace_object(object, trace_reference, NULL);
 }
